requirements.txt
Django==4.2
psycopg2-binary==2.9.6
django-environ==0.11.2
petl==1.7.15
black==24.4.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
freezegun==1.5.1
requests==2.32.2
django-filter==24.2